pci scan and enhanched loggging

+++ include/io.h
@@ -31,7 +31,7 @@ static inline uint8_t inw(uint16_t port)
                    : "Nd"(port) );
     return ret;
 }
-static inline uint8_t inl(uint16_t port)
+static inline uint32_t inl(uint16_t port)
 {
     uint32_t ret;
     asm volatile ( "inl %1, %0"
